<UserBlogsListInDashboard /> refactored + deleteBlog() & blogs in use-database

+++ src/components/UserBlogsListInDashboard.js
@@ -1,51 +1,35 @@
 import { useState, useEffect } from "react";
-import { database } from "../firebase";
-import { ref, remove, onValue } from "firebase/database";
+import { useDatabase } from "../hooks/use-database";
+import { useAuth } from "../hooks/use-auth";
 import { Link, Switch, Route, useRouteMatch } from "react-router-dom";
 import BlogEditionForm from "./BlogEditionForm";
 
-export default function UserBlogsListInDashboard({ userId, setBlogKeyForNewArticle, setBlogTitleForNewArticle }) {
+export default function UserBlogsListInDashboard({ setBlogKeyForNewArticle, setBlogTitleForNewArticle }) {
 
     let {path, url} = useRouteMatch();
 
-    const [userBlogsList, setUserBlogsList] = useState(null);
+    const { blogs, deleteBlog } = useDatabase();
+    const { user } = useAuth();
+
+    const [userBlogs, setUserBlogs] = useState(null);
+
+    useEffect(() => {
+        if (blogs && user) {
+            console.log(blogs);
+            const fetchedBlogs = Object.entries(blogs);
+            const currentUserBLogs = fetchedBlogs.filter(blog => blog[1].userId === user.uid)
+            setUserBlogs(currentUserBLogs);
+        } else {
+            console.log("there are no user and blogs...")
+        }
+    }, [blogs, user]);
 
     const [currentBlogKey, setCurrentBlogKey] = useState(null);
     const [currentBlogLink, setCurrentBlogLink] = useState(null);
 
-    function fetchUserBlogsList(userId) {
-        const userBlogsRef = ref(database, 'users/' + userId + '/blogs');
-        onValue(userBlogsRef, (snapshot) => {
-            if (snapshot) {
-                const fetchedUserBlogsList = snapshot.val();        
-                setUserBlogsList(fetchedUserBlogsList ? Object.entries(fetchedUserBlogsList) : null);
-                console.log("User blogs list fetched to dashboard", fetchedUserBlogsList);
-            }
-        });
-    }
-
-    function deleteBlog(blogKey) {
-        // eslint-disable-next-line no-restricted-globals
-        const wantToDeleteBlog = confirm("Are you sure, you want to delete this blog & all articles from this blog forever? There's no turning back... Delete blog?");
-        if (wantToDeleteBlog) {
-            remove(ref(database, 'users/' + userId + '/blogs/' + blogKey)).then(() => {
-                console.log("blog was deleted");
-                fetchUserBlogsList(userId);
-            }).catch((error) => {
-                // An error ocurred
-                console.log(error.message);
-            });
-            remove(ref(database, 'blogs/' + blogKey));
-        }
-    }
-
     function convertBlogTitleIntoLink(blogTitle) {
         return ("/" + blogTitle.replace(/ /g, "-").toLowerCase());
     }
-
-    useEffect(() => {
-        fetchUserBlogsList(userId);
-    }, [userId]);
 
     return (
         <Switch>
@@ -57,8 +41,8 @@ export default function UserBlogsListInDashboard({ userId, setBlogKeyForNewArtic
                         className="btn btn-info d-block my-3"
                     >Create new blog</Link>
                     {
-                        userBlogsList && userBlogsList.length ?
-                            userBlogsList.map((blog) =>
+                        userBlogs && userBlogs.length ?
+                            userBlogs.map((blog) =>
                                 <div className="container" key={blog[0]}>
                                     <div className="row">
                                         <div className="col">
@@ -87,7 +71,11 @@ export default function UserBlogsListInDashboard({ userId, setBlogKeyForNewArtic
                                                 className="text-danger"
                                                 to="/dashboard/user-blogs"
                                                 onClick={() => {
-                                                    deleteBlog(blog[0]);
+                                                    // eslint-disable-next-line no-restricted-globals
+                                                    const wantToDeleteBlog = confirm("Are you sure, you want to delete this blog & all articles from this blog forever? There's no turning back... Delete blog?");
+                                                    if (wantToDeleteBlog) {
+                                                        deleteBlog(blog[0]);
+                                                    }
                                                 }}
                                             >
                                                 <i className="bi bi-trash" />
@@ -108,9 +96,7 @@ export default function UserBlogsListInDashboard({ userId, setBlogKeyForNewArtic
             </Route>
             <Route path={currentBlogLink}>
                 <BlogEditionForm
-                    userId={userId}
                     blogKey={currentBlogKey}
-                    deleteBlog={deleteBlog}
                 />
             </Route>
         </Switch>
